pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-gmail"
version = "0.1.0"
description = "Multi-tenant Gmail MCP server with OAuth hardening. Drop in for any MCP-compatible client. FastAPI, Postgres, Railway-ready."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pyjwt[crypto]>=2.9",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",
    "cryptography>=42",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov=mcp_gmail --cov-report=term --cov-fail-under=80"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["mcp_gmail"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false
